replace 20k LM with 70k pruned LM, fix trie LM scoring, fix tests, deal with line-endings
git-svn-id: svn+ssh://svn.code.sf.net/p/cmusphinx/code/trunk/sphinx4@13074 94700074-3cef-4d97-a70e-9c8c206c02f5

+++ sphinx4-core/src/main/java/edu/cmu/sphinx/linguist/language/ngram/trie/NgramTrie.java
@@ -125,7 +125,7 @@ public class NgramTrie {
      * Calculates pivot for binary search
      */
     private int calculatePivot(int offset, int range, int width) {
-    	return (offset * width) / (range + 1);
+    	return (int)(((long)offset * width) / (range + 1));
     }
 
     /**
